requirements.txt
Django==4.2.6
django-cors-headers==4.3.0 
django-filter==23.3        
djangorestframework==3.14.0
Markdown==3.5